# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project OpenApi Documentation",
        "description": "This page describes the API documentation of the IPMEDTH project for Revalidatie Friesland",
        "contact": {
            "email": "info@ipmedth.nl"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://ipmedth.nl/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/measurements": {
            "get": {
                "tags": [
                    "Measurements"
                ],
                "summary": "Get all measurements",
                "description": "Returns all measurements",
                "operationId": "GetMeasurements",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Measurements"
                ],
                "summary": "Create a new measurement",
                "description": "Returns a new measurement",
                "operationId": "CreateMeasurement",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreMeasurementRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/measurements/{id}": {
            "get": {
                "tags": [
                    "Measurements"
                ],
                "summary": "Get a measurement",
                "description": "Returns a measurement. Get the related image by calling the image endpoint: https://ipmedth.nl/images/measurements/{filename}",
                "operationId": "GetMeasurement",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Measurement id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Measurements"
                ],
                "summary": "Update a measurement",
                "description": "Returns updated measurement",
                "operationId": "UpdateMeasurement",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Measurement id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateMeasurementRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Measurements"
                ],
                "summary": "Delete a measurement",
                "description": "Returns deleted measurement",
                "operationId": "DeleteMeasurement",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Measurement id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreMeasurementRequest": {
                "required": [
                    "session_id",
                    "hand_view",
                    "hand_type",
                    "hand_score",
                    "finger_thumb",
                    "finger_index",
                    "finger_middle",
                    "finger_ring",
                    "finger_pink",
                    "wrist"
                ],
                "properties": {
                    "session_id": {
                        "description": "Session id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "hand_type": {
                        "description": "The hand that was photographed",
                        "type": "string",
                        "example": "left"
                    },
                    "hand_view": {
                        "description": "What angle was the photo taken from",
                        "type": "string",
                        "example": "thumb_side"
                    },
                    "hand_score": {
                        "description": "The score of the hand",
                        "type": "number",
                        "format": "float",
                        "example": 0.85
                    },
                    "finger_thumb": {
                        "description": "Thumb",
                        "type": "string",
                        "example": "{}"
                    },
                    "finger_index": {
                        "description": "Index finger",
                        "type": "string",
                        "example": "{}"
                    },
                    "finger_middle": {
                        "description": "Middle finger",
                        "type": "string",
                        "example": "{}"
                    },
                    "finger_ring": {
                        "description": "Ring finger",
                        "type": "string",
                        "example": "{}"
                    },
                    "finger_pink": {
                        "description": "Little finger",
                        "type": "string",
                        "example": "{}"
                    },
                    "wrist": {
                        "description": "Wrist",
                        "type": "string",
                        "example": "{}"
                    },
                    "image": {
                        "description": "Photo of the hand the was made",
                        "type": "string",
                        "format": "binary",
                        "example": "file"
                    }
                },
                "type": "object"
            },
...
            "UpdateMeasurementRequest": {
                "required": [
                    "session_id",
                    "hand_view",
                    "hand_type",
                    "hand_score",
                    "finger_thumb",
                    "finger_index",
                    "finger_middle",
                    "finger_ring",
                    "finger_pink",
                    "wrist"
                ],
                "properties": {
                    "session_id": {
                        "description": "Session id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "hand_type": {
                        "description": "The hand that was photographed",
                        "type": "string",
                        "example": "left"
                    },
                    "hand_view": {
                        "description": "What angle was the photo taken from",
                        "type": "string",
                        "example": "thumb_side"
                    },
                    "hand_score": {
                        "description": "The score of the hand",
                        "type": "number",
                        "format": "float",
                        "example": 0.85
                    },
                    "finger_thumb": {
                        "description": "Thumb",
                        "type": "string",
                        "example": "{}"
                    },
                    "finger_index": {
                        "description": "Index finger",
                        "type": "string",
                        "example": "{}"
                    },
                    "finger_middle": {
                        "description": "Middle finger",
                        "type": "string",
                        "example": "{}"
                    },
                    "finger_ring": {
                        "description": "Ring finger",
                        "type": "string",
                        "example": "{}"
                    },
                    "finger_pink": {
                        "description": "Little finger",
                        "type": "string",
                        "example": "{}"
                    },
                    "wrist": {
                        "description": "Wrist",
                        "type": "string",
                        "example": "{}"
                    },
                    "image": {
                        "description": "Photo of the hand the was made",
                        "type": "string",
                        "format": "binary",
                        "example": "file"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}